Refactored the 'start of line' and 'end of line' commands.
